requirements.txt
sqlmodel
python-dotenv
psycopg2-binary
fastapi>=0.95
uvicorn[standard]
requests
httpx
passlib[bcrypt]
python-jose[cryptography]
python-multipart
alembic